Fix merge

+++ App/src/domain/entities/Discount.cs
@@ -1,6 +1,6 @@
 ﻿namespace domain.entities
 {
-    public delegate decimal DiscountStrategy(Cart cart);
+    public delegate double DiscountStrategy(Cart cart);
     public class DiscountStrategyProvider
     {
         public DiscountType DiscountType { get; set; }
@@ -11,12 +11,12 @@
             return cart => { 
                 switch (DiscountType)
                 {
-                    case DiscountType.Percentage: 
-                        decimal discountPercentage = cart.CalculateTotal() * DiscountValue / 100m;
-                        decimal discountTotal = cart.CalculateTotal() - discountPercentage;
+                    case DiscountType.Percentage:
+                        double discountPercentage = cart.CalculateTotal() * DiscountValue / 100;
+                        double discountTotal = cart.CalculateTotal() - discountPercentage;
                         return discountTotal < 1 ? 1 : discountTotal;
                     case DiscountType.FixedAmount:
-                        decimal discountFixed = cart.CalculateTotal() - DiscountValue;
+                        double discountFixed = cart.CalculateTotal() - DiscountValue;
                         return discountFixed < 1 ? 1 : discountFixed; 
                     default:
                         return cart.CalculateTotal();
